+++ lib/rley/parser/gfg_parsing.rb
@@ -1,4 +1,5 @@
 require_relative 'gfg_chart'
+require_relative 'error_reason'
 require_relative 'parse_entry_tracker'
 require_relative 'parse_forest_factory'
 
@@ -15,22 +16,21 @@ module Rley # This module is used as a namespace
       # The sequence of input token to parse
       attr_reader(:tokens)
 
-      # A Hash with pairs of the form: 
+      # A Hash with pairs of the form:
       # parse entry => [ antecedent parse entries ]
       # It associates to a every parse entry its antecedent(s), that is,
-      # the parse entry/ies that causes the key parse entry to be created 
+      # the parse entry/ies that causes the key parse entry to be created
       # with one the gfg rules
       attr_reader(:antecedence)
 
-      # @param aTracer [ParseTracer] An object that traces the parsing.
-      # The possible values are:
-      # 0: No trace output (default case)
-      # 1: Show trace of scanning and completion rules
-      # 2: Same as of 1 with the addition of the prediction rules
-      def initialize(theGFG, theTokens, aTracer)
+      # The reason of a parse failure
+      attr_reader(:failure_reason)
+
+
+      def initialize(theGFG, theTokens)
         @gf_graph = theGFG
         @tokens = theTokens.dup
-        @chart = GFGChart.new(tokens.size, gf_graph, aTracer)
+        @chart = GFGChart.new(tokens.size, gf_graph)
         @antecedence = Hash.new { |hash, key| hash[key] = [] }
         antecedence[chart[0].first]
       end
@@ -45,7 +45,7 @@ module Rley # This module is used as a namespace
         next_symbol = anEntry.next_symbol
         start_vertex = gf_graph.start_vertex_for[next_symbol]
         pos = aPosition
-        apply_rule(anEntry, start_vertex, pos, pos, :call_rule)        
+        apply_rule(anEntry, start_vertex, pos, pos, :call_rule)
       end
 
       # Let the current sigma set be the ith parse entry set.
@@ -65,7 +65,7 @@ module Rley # This module is used as a namespace
       end
 
       # This method must be invoked when an entry is added to a parse entry set
-      # and is of the form [B => γ ., k] (the dot is at the end of the 
+      # and is of the form [B => γ ., k] (the dot is at the end of the
       # production. Then entry [B., k] is added to the current entry set.
       # Gist: for an entry corresponding to a reduced production, add an entry
       # for each exit edge in the graph.
@@ -96,11 +96,12 @@ module Rley # This module is used as a namespace
       end
 
       # Given that the terminal t is at the specified position,
-      #   Locate all entries in the current sigma set that expect t: 
+      #   Locate all entries in the current sigma set that expect t:
       #     [A => α . t γ, i]
       #     and allow them to cross the edge, adding the node on the back side
       #     of the edge as an entry to the next sigma set:
       #       add an entry to the next sigma set [A => α t . γ, i]
+      # returns true if next token matches the expectations, false otherwise.
       def scan_rule(aPosition)
         terminal = tokens[aPosition].terminal
 
@@ -108,7 +109,10 @@ module Rley # This module is used as a namespace
         expecting_term = chart[aPosition].entries4term(terminal)
 
         # ... if the terminal isn't expected then we have an error
-        handle_error(aPosition) if expecting_term.empty?
+        if expecting_term.empty?
+          unexpected_token(aPosition)
+          return false
+        end
 
         expecting_term.each do |ntry|
           # Get the vertices after the expected terminal
@@ -119,6 +123,8 @@ module Rley # This module is used as a namespace
             apply_rule(ntry, vertex_after_terminal, origin, pos, :scan_rule)
           end
         end
+        
+        return true
       end
 
 
@@ -136,7 +142,7 @@ module Rley # This module is used as a namespace
       end
 
       # Factory method. Builds a ParseForest from the parse result.
-      # @return [ParseForest]      
+      # @return [ParseForest]
       def parse_forest()
         factory = ParseForestFactory.new(self)
 
@@ -148,7 +154,7 @@ module Rley # This module is used as a namespace
       # with origin equal to zero.
       def initial_entry()
         return chart.initial_entry
-      end      
+      end
 
       # Retrieve the accepting parse entry that represents
       # a complete, successful parse
@@ -158,25 +164,43 @@ module Rley # This module is used as a namespace
         return chart.accepting_entry
       end
 
+      # Mark the parse as erroneous
+      def faulty(aReason)
+        @failure_reason = aReason
+      end
+
+      # A notification that the parsing reached an end
+      def done
+        unless self.success? || self.failure_reason
+          # Parse not successful and no reason identified
+          # Assuming that parse failed because of a premature end
+          premature_end
+        end
+      end
+
       private
 
-      # Raise an exception to indicate a syntax error.
-      def handle_error(aPosition)
-        # Retrieve the actual token
-        actual = tokens[aPosition].terminal
-        lexeme_at_pos = tokens[aPosition].lexeme
+      # Parse error detected: all input tokens were consumed and
+      # the parser didn't detect syntax error meanwhile but
+      # could not reach the accepting state.
+      def premature_end
+        token_pos = tokens.size # One-based!
+        last_token = tokens[-1]
+        entry_set = chart.sets[tokens.size]
+        expected = entry_set.expected_terminals
 
+        reason = PrematureInputEnd.new(token_pos - 1, last_token, expected)
+        faulty(reason)
+      end
+
+      # Parse error detected: input token doesn't match
+      # the expectations set by grammar rules
+      def unexpected_token(aPosition)
+        unexpected = tokens[aPosition]
         expected = chart.sets[aPosition].expected_terminals
-        term_names = expected.map(&:name)
-        err_msg = "Syntax error at or near token #{aPosition + 1}"
-        err_msg << ">>>#{lexeme_at_pos}<<<:\nExpected "
-        err_msg << if expected.size > 1
-                     "one of: ['#{term_names.join("', '")}'],"
-                   else
-                     ": #{term_names[0]},"
-                   end
-        err_msg << " found a '#{actual.name}'"
-        raise StandardError, err_msg + ' instead.'
+
+        reason = UnexpectedToken.new(aPosition, unexpected, expected)
+        faulty(reason)
       end
 
       def apply_rule(antecedentEntry, aVertex, anOrigin, aPosition, aRuleId)