+++ lib/rley/rley_error.rb
@@ -1,0 +1,10 @@
+# File: exceptions.rb
+
+module Rley # Module used as a namespace
+  # @abstract
+  # Base class for any exception explicitly raised by Rley code.
+  class RleyError < StandardError
+  end # class
+end # module
+
+# End of file